requirements.txt
filelock==3.17.0
fsspec==2025.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
pillow==11.1.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
